Add files via upload
This code is for pre-existing images (for example, something pulled from wikipedia). It uses build-in functions from the Images class in the Pillow PIL package to execute isometries on the image/tessellation and take snapshots of the lens view. It then converts the images into numpy arrays and computes the persistent homologies. 

Notes: currently, the code throws an error when the lens falls off of the page, but doesn't have an actual algorithm for knowing when this will happen. For the future, figure out exactly how much you can translate after a rotation of the plane before falling off